Corrigindo comentários
