+++ src/components/ScheduleRoom/SectionOfInput/RepeatDateButton/lib.ts
@@ -1,5 +1,1 @@
-export const buttonsRepeat = [
-  { label: "daily", value: "daily" },
-  { label: "weekly", value: "weekly" },
-  { label: "no", value: "no" },
-];
+export const buttonsRepeat = ["daily", "weekly", "no"];
